+++ terminology/src/main/java/org/ehrbase/openehr/sdk/terminology/openehr/implementation/TerminologySourceFactory.java
@@ -29,7 +29,8 @@ public class TerminologySourceFactory {
     private static final String OPENEHR_TERMINOLOGY_EN = "openehr/en/openehr_terminology.xml";
     private static final String OPENEHR_TERMINOLOGY_JA = "openehr/ja/openehr_terminology.xml";
     private static final String OPENEHR_TERMINOLOGY_PT = "openehr/pt/openehr_terminology.xml";
-    private static final String EXTERNAL_TERMINOLOGIES = "openehr_external_terminologies.xml";
+    private static final String OPENEHR_TERMINOLOGY_ES = "openehr/es/openehr_terminology.xml";
+    private static final String EXTERNAL_TERMINOLOGIES = "openehr/openehr_external_terminologies.xml";
 
     /**
      * Gets an instance of openEHR terminology source
@@ -44,6 +45,8 @@ public class TerminologySourceFactory {
                 return XMLTerminologySource.getInstance(OPENEHR_TERMINOLOGY_JA);
             case "pt":
                 return XMLTerminologySource.getInstance(OPENEHR_TERMINOLOGY_PT);
+            case "es":
+                return XMLTerminologySource.getInstance(OPENEHR_TERMINOLOGY_ES);
             default:
                 return XMLTerminologySource.getInstance(OPENEHR_TERMINOLOGY_EN);
         }